Address review feedback on #15165 HTMX partial swap fix
- Add `__deepcopy__` to `AttrSelectMixin` so each per-instance widget copy
  owns its `descriptions` dict. Django's `ChoiceWidget.__deepcopy__` copies
  attrs and choices but not descriptions, so copies otherwise shared one
  dict with the class-level widget: latent cross-request state that adding
  the mixin to `HTMXSelect` widened.
- Remove the dead `mode` entries from `Meta.labels` on both interface forms.
  Django ignores `Meta.labels` for explicitly declared fields, the same
  reason it ignored `Meta.widgets`; the label is set on the declared field.
- Assert an HTMX verb (`hx-get`) in the partial-swap tests. `hx-target` and
  `hx-select` alone issue no request, which was the reported symptom.
- Resolve the swap subwidget by type (`isinstance(w, HTMXSelect)`) rather
  than a hardcoded index, so a change in MultiWidget ordering cannot make
  the full-form assertion pass vacuously.
- Move `HTMXPartialSwapRenderingTestCase` to `netbox/tests/`. It imports
  forms from eight apps; `utilities` is the model-free bottom layer and
  should not fail when an app's forms break.
